Fixed bug #157

+++ public/Import-AdfFromFolder.ps1
@@ -61,12 +61,15 @@ function Import-AdfFromFolder {
     # A workaround of Microsoft's bug - no 'properties' in ManagedVirtualNetwork object
     if ($adf.ManagedVirtualNetwork.Count -eq 1) {
         $o = $adf.ManagedVirtualNetwork[0]
-        Set-StrictMode -Version 1.0
-        Add-ObjectProperty    -obj $o.Body -path 'properties.preventDataExfiltration' -value $false -ErrorAction 'Continue'
-        Remove-ObjectProperty -obj $o.Body -path 'properties.preventDataExfiltration'               -ErrorAction 'Continue'
-        Write-Verbose 'Workaround: Added empty "properties" node to ManagedVirtualNetwork object.'
-        $f = (Save-AdfObjectAsFile -obj $o)
-        $o.FileName = $f
+        if ($o.Body.PSobject.Properties.Name -notcontains "properties")
+        {
+            Write-Verbose 'Workaround: Addeding empty "properties" node to ManagedVirtualNetwork object...'
+            Set-StrictMode -Version 1.0
+            Add-ObjectProperty    -obj $o.Body -path 'properties.preventDataExfiltration' -value $false -ErrorAction 'Continue'
+            Remove-ObjectProperty -obj $o.Body -path 'properties.preventDataExfiltration'               -ErrorAction 'Continue'
+            $f = (Save-AdfObjectAsFile -obj $o)
+            $o.FileName = $f
+        }
     }
 
     Write-Debug "END: Import-AdfFromFolder()"
